+++ ChangeLog.md
@@ -4,6 +4,11 @@ T-Regx Changelog
 Incoming
 --------
 
+* Soon
+
+Added in 0.37.1
+---------------
+
 * Features
     * Add `PatternList.count()`, which allows to count occurrances of many patterns in a subject.
 * Bug fixes